+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Flatten `Buffer | Uint8Array` types to base-only `Uint8Array`
+
+
 ## 12.0.1 Apr 22, 2023
 
 Contributed: